% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_feature}
\alias{aa_feature}
\title{amino acid feature index, scaled value from -1 to 1}
\format{
\subsection{\code{aa_feature}}{

A data frame with 20 rows and 5 columns:
\describe{
\item{hydrophilicity}{Hydrophobicity (Prabhakaran, 1990)}
\item{polarity}{Polarity (Grantham, 1974)}
\item{charge}{Net charge (Klein et al., 1984)}
\item{volume}{Side chain volume (Krigbaum-Komoriya, 1979)}
}
}
}
\source{
adjusted from \href{https://www.genome.jp/aaindex/}{aaindex}
}
\usage{
aa_feature
}
\description{
amino acid feature index, scaled value from -1 to 1
}
\keyword{datasets}
