% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aa_info}
\alias{aa_info}
\title{amino acid features}
\usage{
aa_info(x, raw = FALSE)
}
\arguments{
\item{x}{aa}

\item{raw}{return feature raw index instead of scaled index}
}
\value{
data.frame
}
\description{
amino acid features
}
\examples{
aa_info("C")
}
