% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAMSA-class.R
\docType{class}
\name{BiologyAAMSA-class}
\alias{BiologyAAMSA-class}
\title{a S4 class to operate aligned AA sequence set}
\description{
a S4 class to operate aligned AA sequence set
}
\section{Slots}{

\describe{
\item{\code{AA}}{AAStringSet.}

\item{\code{params}}{MSA params}

\item{\code{consAA}}{consensus AAString}

\item{\code{consAAfreq}}{AA consensus frequence}
}}

