% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAMSA-class.R
\name{BiologyAAMSA}
\alias{BiologyAAMSA}
\title{create BiologyAAMSA object}
\usage{
BiologyAAMSA(x, skip_align = FALSE, order = "input", method = "Muscle")
}
\arguments{
\item{x}{AAStringSet}

\item{skip_align}{skip align or not}

\item{order}{sequences order, \code{input} to keep input order, \code{aligned} to use
aligned order}

\item{method}{msa method, \code{ClustalOmega|Muscle|ClustalW}.}
}
\value{
BiologyAAMSA
}
\description{
create BiologyAAMSA object
}
\examples{
AA <- c('MQVNPTE', 'MQVTE', 'MQVTV')

aln <- BiologyAAMSA(AA)

AA(aln)

consAA(aln)

aln[[2]]

}
