% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAmut-class.R
\docType{class}
\name{BiologyAAmut-class}
\alias{BiologyAAmut-class}
\title{a S4 class to operate AA mutation}
\description{
a S4 class to operate AA mutation
}
\section{Slots}{

\describe{
\item{\code{mut}}{AA mutations}

\item{\code{site}}{the sites of mutations}
}}

