% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAMut-class.R
\name{BiologyAAmut}
\alias{BiologyAAmut}
\title{create BiologyAAmut object}
\usage{
BiologyAAmut(mut)
}
\arguments{
\item{mut}{vector of mutations}
}
\value{
BiologyAAmut object
}
\description{
create BiologyAAmut object
}
\examples{
mut <- BiologyAAmut(c("D123G", "D6F", "C878C", "D123G", "C878T"))

mut

sort(mut)

unique(sort(mut))

unique(sort(mut), bysite = TRUE)

}
