% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAmutSet-class.R
\docType{class}
\name{BiologyAAmutSet-class}
\alias{BiologyAAmutSet-class}
\title{a S4 class to operate AA mutation set}
\description{
a S4 class to operate AA mutation set
}
\section{Slots}{

\describe{
\item{\code{muts}}{list of AA mutations}
}}

