% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAmutSet-class.R
\name{BiologyAAmutSet}
\alias{BiologyAAmutSet}
\title{create BiologyAAmutSet object}
\usage{
BiologyAAmutSet(mut_list)
}
\arguments{
\item{mut_list}{list of AA mutations}
}
\value{
\code{BiologyAAmutSet} object
}
\description{
create BiologyAAmutSet object
}
\examples{
mut_list <- list(
  mut1 = c("D123G", "D6F", "C878C", "C878T"),
  mut2 = c("D123G", "D6F"),
  mut3 = c("C878C", "D123G", "C878T")
)

muts <- BiologyAAmutSet(mut_list)

}
