% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeq-class.R
\docType{class}
\name{BiologySeq-class}
\alias{BiologySeq-class}
\title{a S4 class to operate DNA together with AA sequence}
\description{
a S4 class to operate DNA together with AA sequence
}
\section{Slots}{

\describe{
\item{\code{DNA}}{DNAString.}

\item{\code{AA}}{AAString.}
}}

