% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeq-class.R
\name{BiologySeq}
\alias{BiologySeq}
\title{create BiologySeq object}
\usage{
BiologySeq(DNA, corr_gaps = FALSE)
}
\arguments{
\item{DNA}{character of DNA letters}

\item{corr_gaps}{see \code{?DNA_gaps_corr}}
}
\value{
BiologySeq object
}
\description{
create BiologySeq object
}
\examples{

bs <- BiologySeq('ATGAAA---CCCTTTGGG---')

DNA(bs)

AA(bs)

}
