% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeqMSA-class.R
\docType{class}
\name{BiologySeqMSA-class}
\alias{BiologySeqMSA-class}
\title{a S4 class to operate aligned DNA together with AA sequence set}
\description{
a S4 class to operate aligned DNA together with AA sequence set
}
\section{Slots}{

\describe{
\item{\code{params}}{MSA params}

\item{\code{consDNA}}{consensus DNAString}

\item{\code{consDNAfreq}}{DNA consensus frequency}

\item{\code{consAA}}{consensus AAString}

\item{\code{consAAfreq}}{AA consensus frequence}

\item{\code{consSeq}}{consensus BiologySeq}

\item{\code{AAnumbering}}{numbering}
}}

