% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeqSet-class.R
\docType{class}
\name{BiologySeqSet-class}
\alias{BiologySeqSet-class}
\title{a S4 class to operate DNA together with AA sequence set}
\description{
a S4 class to operate DNA together with AA sequence set
}
\section{Slots}{

\describe{
\item{\code{DNA}}{DNAStringSet.}

\item{\code{AA}}{AAStringSet.}
}}

