% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeqSet-class.R
\name{BiologySeqSet}
\alias{BiologySeqSet}
\title{create BiologySeqSet object}
\usage{
BiologySeqSet(DNA, corr_gaps = FALSE)
}
\arguments{
\item{DNA}{character vector of DNA letters}

\item{corr_gaps}{see \code{?DNA_gaps_corr}}
}
\value{
BiologySeqSet object
}
\description{
create BiologySeqSet object
}
\examples{

bs <- BiologySeqSet(c('ATGAAA---', 'ATG---AAA'))

DNA(bs)

AA(bs)

}
