% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomuts-package.R
\docType{package}
\name{biomuts-package}
\alias{biomuts}
\alias{biomuts-package}
\title{biomuts: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com}

}
\keyword{internal}
