% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_AAmut}
\alias{call_AAmut}
\title{call AA mutations from two AAString objects}
\usage{
call_AAmut(query, ref)
}
\arguments{
\item{query}{Biostrings::AAString}

\item{ref}{Biostrings::AAString}
}
\value{
vector of AA mutations
}
\description{
call AA mutations from two AAString objects
}
\examples{

call_AAmut(
  Biostrings::AAString("MQVNPTE"),
  Biostrings::AAString("MQVCTTE")
)
}
