% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyAAMSA-class.R
\name{call_AAmutSet,BiologyAAMSA-method}
\alias{call_AAmutSet,BiologyAAMSA-method}
\title{call AA mutations from \code{BiologyAAMSA} object}
\usage{
\S4method{call_AAmutSet}{BiologyAAMSA}(querys, ref = "consensus", rm_ref = TRUE)
}
\arguments{
\item{querys}{BiologyAAMSA object}

\item{ref}{\code{consensus}, or the name of sequence}
}
\value{
\code{BiologyAAmutSet} object
}
\description{
call AA mutations from \code{BiologyAAMSA} object
}
