% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeqMSA-class.R
\name{call_AAmutSet,BiologySeqMSA-method}
\alias{call_AAmutSet,BiologySeqMSA-method}
\title{call AA mutations from \code{BiologySeqMSA} object}
\usage{
\S4method{call_AAmutSet}{BiologySeqMSA}(querys, ref = "consensus", rm_ref = TRUE)
}
\arguments{
\item{querys}{BiologySeqMSA object}

\item{ref}{\code{consensus}, or the name of sequence}
}
\value{
\code{BiologyAAmutSet} object
}
\description{
call AA mutations from \code{BiologySeqMSA} object
}
