% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_AAmutSet}
\alias{call_AAmutSet}
\title{call AA mutations from BiologySeqMSA object}
\usage{
call_AAmutSet(querys, ref = "consensus", rm_ref = TRUE)
}
\arguments{
\item{querys}{BiologySeqMSA object}

\item{ref}{\code{consensus}, or the name of sequence}
}
\value{
BiologyAAmutSet
}
\description{
call AA mutations from BiologySeqMSA object
}
\examples{
bss <- BiologySeqSet(
  c("ATGCAGGTAAACCCTACTGAG", "ATGCAGGTTACTGAG", "ATGCAGGTAACTGTG")
)
alnbs <- BiologySeqMSA(bss)

call_AAmutSet(alnbs)

call_AAmutSet(alnbs, ref=3)

}
