% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_aa}
\alias{compare_aa}
\title{compare feature index between AAs}
\usage{
compare_aa(query, ref, raw = FALSE)
}
\arguments{
\item{query}{query aa}

\item{ref}{ref aa}
}
\value{
data.frame
}
\description{
compare feature index between AAs
}
\examples{
compare_aa(c('A', 'T'), 'C')
}
