% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dna2aa}
\alias{dna2aa}
\title{translate DNA to AA (gaps allowed)}
\usage{
dna2aa(x)
}
\arguments{
\item{x}{character of DNA letters, or Biostrings::DNAString object}
}
\value{
character of AA letters
}
\description{
translate DNA to AA (gaps allowed)
}
\examples{
dna2aa("ATG---AAA")
}
