% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DNA_gaps_corr}
\alias{DNA_gaps_corr}
\title{correct the gaps location to fit codon frames}
\usage{
DNA_gaps_corr(x, dismiss_no3gaps = FALSE)
}
\arguments{
\item{x}{character of DNA letters}

\item{dismiss_no3gaps}{ignore gap widths that are not multiples of 3}
}
\value{
character of DNA letters
}
\description{
correct the gaps location to fit codon frames
}
\examples{
DNA_gaps_corr("AT---G")

}
