% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_mut}
\alias{number_mut}
\title{number the mutations}
\usage{
number_mut(x, numbering)
}
\arguments{
\item{x}{character vector of mutations}

\item{numbering}{vector of numbering}
}
\value{
character vector of mutations
}
\description{
number the mutations
}
\examples{

number_mut(c("A12T", "C18P"), c("12" = "12", "18" = "18"))

}
