% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologySeqMSA-class.R
\name{BiologySeqMSA}
\alias{BiologySeqMSA}
\title{create BiologySeqMSA object}
\usage{
BiologySeqMSA(x, corr_gaps = TRUE, order = "input", method = "ClustalOmega")
}
\arguments{
\item{x}{BiologySeqSet}

\item{corr_gaps}{correct gaps codon frame after MSA, \code{TRUE} as default.
see \code{?DNA_gaps_corr}}

\item{order}{sequences order, \code{input} to keep input order, \code{aligned} to use
aligned order}

\item{method}{msa method, \code{ClustalOmega|Muscle|ClustalW}.}
}
\value{
BiologySeqMSA object
}
\description{
create BiologySeqMSA object
}
\examples{

bss <- BiologySeqSet(c("ATGAAA---", "ATG---AAA"))

BiologySeqMSA(bss)

}
