% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_mut}
\alias{call_mut}
\title{call DNA and AA mutations from two BiologySeq objects}
\usage{
call_mut(query, ref)
}
\arguments{
\item{query}{BiologySeq}

\item{ref}{BiologySeq}
}
\value{
list
}
\description{
call DNA and AA mutations from two BiologySeq objects
}
\examples{

ref <- BiologySeq("CTTCCCTTC")
query <- BiologySeq("ATGCCCTTT")
call_mut(query, ref)

}
