% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_muts}
\alias{count_muts}
\title{count AA mutations from \code{BiologyAAmutSet} object}
\usage{
count_muts(muts, bysite = FALSE, use_numbering = FALSE)
}
\arguments{
\item{muts}{\code{BiologyAAmutSet} object}

\item{bysite}{merge the mutations at the same site}

\item{use_numbering}{use numbering or not, \code{FALSE} as default}
}
\value{
tibble
}
\description{
count AA mutations from \code{BiologyAAmutSet} object
}
\examples{
mut_list <- list(
  mut1 = c("D123G", "D6F", "C878C", "C878T"),
  mut2 = c("D123G", "D6F"),
  mut3 = c("C878C", "D123G", "C878T")
)

muts <- BiologyAAmutSet(mut_list)

count_muts(muts)

}
