% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutset_compare.R
\name{mutset_compare}
\alias{mutset_compare}
\title{compare two \code{BiologyAAmutSet} objects by Chi-Squared test}
\usage{
mutset_compare(x, y, min_count = 2, all_tests = FALSE)
}
\arguments{
\item{x}{BiologyAAmutSet}

\item{y}{BiologyAAmutSet}

\item{min_count}{minimal count of a mutation to perform test}

\item{p_threshold}{0.05 as default}
}
\value{
tibble
}
\description{
compare two \code{BiologyAAmutSet} objects by Chi-Squared test
}
\examples{
BiologyAAmutSet(list("D12C", "D12C"))
}
